make UI more adaptive for different screen sizes
Especially the bulb vector graphic inside the rgb circle scales in a way, so it
falls within the rgb circle for any screen size.
It is achieved by setting the normal size of the bulb vector graphic relative to the
size of the rgb circle png image so that the bulb is inside the rgb circle.
When the bulb and the rgb circle are scaled (because of bigger/smaller screen),
the bulb stays within the rgb circle image.