implement rgb circle class
Represents a rgb circle, which knows what rgb values it has at certain angles.
Each color has a area in the rgb circle.
Each color area has a certain size. The center of a color area is half the size of the whole
color area and contains max values of the respective color.
Around the center are areas where the color values fade away as the angle gets closer to
the start/end of the color area.