change alarm data class
An alarm now has an id, a trigger time, an activated status and a byte
whose bits indicate on which day the alarm is repetitive.
The 7 least significant bits of this byte indicate whether the alarm will be triggered on
the corresponding day of week.
Monday is assigned to the second most significant bit followed by the other week days
in regular order (2nd most significant bit -> Monday, ..., least significant bit -> Sunday).
The most significant bit has no meaning.