set brightness only if user changes brightness
setting the current brightness after the current settings
are loaded will change the progress of the brightness
seekbar, because the current brightness value and the
progress of the seekbar are conntected via data binding.
When this happens the brightness should not be set again
and there should not be a RgbRequest setting the brightness.
This should only happen if the user changes the brightness
via the brightness seekbar.