replace ControllerRepository with RgbRequestRepository
The RgbRequestRepository uses a single instance of RgbRequestService
that works with multiple ip addresses. This allows to have
a changeable number of rgb strips at runtime.
Also the RgbRequestRepository is not only limited to the controller
feature and can be used in the rgb show feature, for example.
The available rgb strips are loaded from the Esp32Microcontrollers whose
ip addresses are stored in the app (ip addresses are currently
hard-coded, later the user should enter or select the ip addresses).