fix bug when updating rgb alarm
Reason of bug: there was an endless loop in DefaultRgbAlarmRepository at
the alarms property because the data in the database changed with each
call to alarmDao.insertOrUpdate. This happened even when no new rgb
alarm was passed to insertOrUpdate(), since OnConflictStrategy.REPLACE
was used, which removes and inserts the already existing rgb alarm
resulting in a change in the database which triggers the mapLatest of
the alarms property.