make ...AddEditFragments survive configuration changes
RgbAlarmAddEditFragment and IpAddressAddEditFragment now survive
configuration changes.
Reason for bug: Before each call to onViewCreated() (e.g. when
configuration change happens) the method for setting the selected
RgbAlarm / IpAddressNamePair was called. This caused the
RgbAlarm / IpAddressNamePair to be loaded as it is stored in the
DataStore / Database, losing the changes made by the user.